Add the ability to precompile assets to the source directory.
This enables using safe mode and the default jekyll on github pages,
as long it `jekyll build` is run once in unsafe mode after files in
the _assets directory are modified.

I'm doing this because `jekyll-assets` is the only plugin I'm using
and this gives me the flexibility to make changes to `_posts` using
the github interface or something.