Fix bootstrap plugin
Actually it just makes simplecov think that all lines are tested.
I believe the best would be to write tests against paths that were added
instead of integrational for bootstrap so that we could check which paths
are added on different version (by stubbing gemspec.version message).

But, I must admit I'm mostly don't care about these out-of-the-box plugins.
And they are subjects of complete removeal in 1.0.0 release.