Introducing proxies and re-introducing image manipulation.
Fixes #160: Use {% img name.ext magick:crop:value %}
Fixes #149: Use {% img name.ext magick:2x %} also 4x, half.
Insur #159: Use a hash of the proxies for the image path.
Fixes #150: Simply because the new syntax is solidified.

This patch introduces the concept of proxies, so that tags can have
proxies that manipulate data, the base included proxy is a magick filter
that allows you to adjust and manipulate your image but any number of
filters can be used.

Because an image can be filtered any number of times each proxy that
changes an image will trigger a new hash resulting in an entirely uniq
image but if you use an image with a filter 20 times it should in theory
continue to cache it and only run the proxies once and hold that cache
until you either change the proxy attributes or clear the cache.