Rogue can cause a segfault sometimes and Jekyll does not exclude.
Something in Jekyll 3.0 changed and now `_assets` is no longer excluded by
default anymore (if it ever was?) and Rogue can segfault sometimes after you
update an asset, making it PTA to develop.  Try to avoid this by making sure our
source folders and cache folders are excluded.  It's not only our addon that can
do this, apparently dozens of other addons can do this too since there were
reports through `jekyll/docker` of this happening and that user did have
`jekyll-assets` installed.

This is not considered a bug in jekyll-assets, it's a bug in Ruby through Jekyll
through Rouge but we are trying to address it and prevent our software from
triggering that segfault that can happen.