Solve this damn asset rewrite problem once and for all.
This rewrites the way we handle used assets entirely by using the manifest to handle all the assets, as long as there is a cache, all the cached assets will get re-written every time by pulling out all the files and adding the used.  This will result in even sub-assets that don't need to be re-processed being written.  Hopefully, if that's not the case, then I'll officially consider this an upstream design problem.  Good riddance to missing assets on rebuild.