Fix #284: Revert our removal of patches for Jekyll::Cleaner.
This reverts and readds our patches to Jekyll::Cleaner.  When we originally removed the patch we had assumed the behavior of "keep_files" but it seems we were wrong.  This reverts that behavior and starts patching obsolete_files again so that we don't lose our assets on a site recompile.