Add support for <img asset> parsing.
This makes it so so you can see assets as you work in Markdown. Instead
of having to see something like `{% asset %}`, it even works in HTML,
though I don't know why you would want that, but if you do, more power
to you I guess.  This is just there so you can abuse it and use it if
you really wish to.